Revert untested changes to linux swap interval control.  It seems the problem the previous changes were
intended to fix that glXGetProcAddress() was not found causing a linking error when loading _graphics.so is because
the build is not linking against libglx.  The fix should be to link against libglx.  Apparently it works without
linking because importing PyOpenGL loads libglx and it is imported before the _graphics.so Python module is imported.
But when offscreen rendering is used, it may be trying to import _graphics.so before PyOpenGL.