Updated to PyPi lxml 4.6.2 from 4.5.2.  GitHub warned 4.5.2 was insecure.
I also removed from Makefile.pip the option that forced compiling lxml from source on macOS.  That was added by Conrad in April 2020 with the comment that it was needed for notarization to succeed.  The 5 lxml shared libraries only depend on a system library so I am not sure why notarization would fail.  Will see if notarization works tonight with this new lxml version.