Tab Bar: Make gray background transparent on macOS
By default, the Qt toolkit tries to make things look native to
macOS. There's nothing wrong with the tabs that we have now, except
that it's somewhat unsightly to have the gray background underneath
both the main part of the toolbar and coming halfway up the tabs.
The only reason it looks awkward is because the rest of the content
is not and should not be contained within the tabs, so we get a very
long, very thin gray box.