prereqs/pips: Add a Pipfile
Missing: pyopengl-accelerate, tinyarray, tables

Removed: suds-jurko (we are removing Opal soon)

We know the importance of virtual environments -- we already use the
Python we bundle with ChimeraX as a virtual environment. Making that
more explicit has many advantages including: using existing Python tooling
to freeze the environment instead of constructing it by hand, and moving
the environment to a location the rest of the ecosystem expects allows,
say, language servers to pick up on our code and assist in developing
and debugging.