build: Use the same directory for all ChimeraX wheels
Since we know at develop time whether a bundle is binary or not, the
separation of binary and pure python pips at build time -- for the
purposes of building only -- seems like unnecessary complexity. This
change also makes it easier to install ChimeraX pips to a virtual
environment, as all existing circular dependencies are resolved by
having all wheels in the same directory.