bundles/dicom: (Trac#6255) Infer time series from files with incomplete metadata
Our time series logic previously relied on NumberOfTemporalPositions
exclusively. This field was assumed to be the same and present in every
file. Based on that, we would read TemporalPositionIdentifier to determine
which files belonged in what series.

Some datasets are missing that metadata but do have a TriggerTime field
that can be used to infer the structure of the time series. This commit
does that inference. We also add a new property to SeriesFile,
inferred_properties, so we can make it clear to users when we have a
DICOM metadata viewer that a field has been constructed and is not original
to the file.