Create a chimerax dir with symlinked bundles, core
This should be the shortest path to being able to run unit and
integration tests on bundles without having to either boot up
ChimeraX or do some wonky virtual environment hacking to get that
job done.

We additionally copy a pre-generated buildinfo.py into this directory
but not the "real" core directory to enable Python to import
chimerax.core successfully.