bundles/ui: Redirect stdout and stderr to log as soon as possible
Basically if the UI bundle exists at all, we should be capturing
stdout and stderr as soon as we're able to. That way, we can use
the Python plain text logger in places where we'd otherwise pass
a session *just for the logger*.

This keeps the same amount of cohesion and decreases coupling,
significantly.