MainWindow: Let bundles configure the central widget
This commit adds an experimental API to configure ChimeraX's central widget:

graphicsArea is added as a method of MainWindow and returns the
widget containing the surface on which we draw graphics. It is also
saved in the hidden attribute _backup_main_view.

main_view is added as a property of MainWindow. It is expected to be
the second widget in the central widget stack. Its setter ensures that
the incoming widget contains the graphics area. If not, it sets
main_view to _backup_main_view and informs the developer-user that
they must put the widget returned by graphicsArea into their new layout.

restore_default_main_view sets the central widget back to the graphics
window's widget. It is expeted to be called on bundle deinitialization.

Bundles must not rely on the GUI module to save the state of an
alternative main_view widget hierarchy. If preservation is desired, they
must provide their own mechanism to ensure it.