Add a subclass of View specialized to the 2x2 view
The existing View class has a very structured draw() method that
sets up the background, then does the drawing, then finalizes rendering
all in one call. No parameters are passed to the draw method that can
easily filter down to individual drawings where we need them to be, so
this View subclass completely takes over rendering so that we do not
have to pass a bunch of parameters through functions that don't need
them.