BundleBuilder: Handle universal2 Python on intel-only Macs
Starting with Python 3.10 or 3.11 (not sure because we skipped 3.10),
only universal2 versions of Python will be distributed. That's a problem
for bundle builder because it used the universal2 platform tag to decide
whether or not it should make fat binaries, and universal2 Python
can run on platforms that aren't compatible with Xcode 12, the first
version that included the libraries needed to make fat binaries.

This change makes it so that we only try to make fat binaries on
platforms that actually _can_ make them: macOS 10.15+