Make ImagePositionPatient the highest priority sort key for DICOM files
The DICOM Z-axis is usually increasing from the first frame to the last
frame, and this is an assumption we make all over the DICOM code. This
change was made after observing a DICOM dataset that had SliceLocation
values that were negative and ImageOrientationPatient values that were
positive, which threw the sorting logic out of whack and resulted in a
model with a decreasing Z axis.

Perhaps in the future the sorting function will need to be even smarter
and test various attributes for whether they are increasing and then pick
the one that is to order the slices by, or deliberately misinterpret a
decreasing attribute if no such increasing attributes are found.