Add a context manager to place ChimeraX's bin directory first in PATH
This is useful for us in case we use a dependency like, say,
openai-whisper, that makes calls to programs that may exist on the
system but we also ship with ChimeraX, and we'd like to use our version
instead of whatever is out there on the system.

Bundles, if they choose to, can use this same technique!