medical_toolbar: Add an icon for the Volume Raycasting rendering mode
This icon was sourced from Wikimedia Commons. The original file can
be found at:

https://commons.wikimedia.org/wiki/File:High_Definition_Volume_Rendering.JPG

At the time of inclusion in ChimeraX, to the best of my knowledge, the
publisher of the image had released it into the public domain as
evidenced by a copyright statement to that effect under the image.