docs(devel): Hide class names in tables of contents
Previously, the class name was prepended to all method names in
tables of contents. For example, if you were looking at the Atomic
module, all methods of AtomicStructure would start with AtomicStructure
in the table of contents, which was highly redundant and took up a lot
of horizontal space. Now, those should be hidden in the method names.
They are indented slightly, which has the same effect that the class
name used to have, if a little more subtly.