fix(blastprotein): If a name has one field take it as the name for onlyBest
Previously setting the onlyBest parameter true made blastprotein attempt
to split hit names so it could find the best chain of a hit to present
to the user. This accidentally also worked for alphafold because
alphafold results look like XXXX_SPECIES, but it broke for esmfold.
Now if the name can't be split we just take the whole name and the
function effectively does nothing.