fix(webservices): More robust checks for restored jobs
Previously restored jobs may not have gotten a next_check, so
the module now assigns "1" if any code attempts to access next_check
and it is none. Badly saved jobs may also have their states set to
RUNNING but also have a start and end time. We now prefer the state
of the job over the implicit check of start and end, but still do
the check as a fallback. Finally, if the client attempts to check the
status of a job that is no longer running we set the state of the job
to FINISHED and warn the user to save their results this time, as the
next time the job is restored Plato may have deleted the results.

Fixes #14980