fix(segmentations): Better saving/restoring of hand/mouse modes
Previously we didn't check whether we were on a platform that supports
VR, or in VR, before attempting to save and restore hand modes, which
caused tracebacks on Mac and Linux. We also didn't guard saving and
restoring mouse modes, so the user could hit restore before save and
set a bunch of mouse modes to 'None' erroneously.

Fixes #15162