fix(segmentations): Remove default format menu entry
In retrospect it's sort of nonsensical to save e.g. DICOM files as
other formats because they may not even be able to adequately describe
the data effectively. It would be best if the tool simply looked at the
segmentation to see what data it was derived from and allowed the user
to save it in that format