feat(segmentations): Have the segmentations bundle track segmentations
Previously, the Segmentations tool tracked segmentations. The tool and
the PlaneViewers would talk to each other if need be, and direct
communication was meticulously set up and deconstructed as view layouts
changed and the segmentation tool opened or closed.

This commit takes one of many steps towards decoupling these objects.
The segmentations tool now uses the segmentations command to add and
remove segmentations. The bundle itself tracks them instead of the
segmentations tool.

There are two new triggers:
SEGMENTATION_ADDED - when this bundle, reacting to the ADD_MODELS
trigger, recognizes a model as a segmentation and adds it to its
Volume -> Segmentations dictionary.
SEGMENTATION_REMOVED - when this bundle, reacting to the REMOVE_MODELS
trigger, recognizes a model as a segmentation and removes it from its
Volume -> Segmentation dictionary.

The Segmentations tool now listens to ADD_MODELS and REMOVE_MODELS to
rebuild its list of segmentations. The PlaneViewers listen to the above
triggers to add and remove overlays representing the open segmentations.

Fixes Trac#10341,Trac#15163