fix(dicom): Add an array attribute to DicomGrid
Usually all segmentations that a bundle opens are going to be opened as
Segmentations with generic ArrayGrids, except DICOM segmentations which
are part of a special DICOM hierarchy and so use DicomGrid. This commit
just adds the attribute that segmentations (the bundle) expects to access
when editing segmentations (the data structure).

Fixes Trac#9526