fix(segmentations): Reformat segmentation names for clarity
Previously the names were something like "segmentation of foo (#x)"
where X was supposed to be the number of extant segmentations for that
volume. Not only did that not work, the number looked suspiciously like
a model number which is confusing. Now the segmentations are enumerated
correctly and the format has changed to "segmentation x of foo" for
clarity.