fix(dicom): Removing and readding a series would not render it
DICOM studies keep a list of open series and their series UIDs so that
if a user opens a file and the series is already open that file is
merged with the existing series and under the correct study instead of
polluting the model menu with redundant patients and studies. The study
previously did not have a way to watch its child models and keep its
list of currently alive, drawn series current, which meant you could
close a series, try to reopen it, and it wouldn't get depicted! Now
studies register handlers for the delete triggers of all child models,
so they know when they're gone and can prepare for when and if they're reopened.