fix(segmentations): Fail silently if 3D seg attempted on no model
Even if you only open a segmentation, you can still open the
segmentations tool. The segmentation won't appear in the menu (it has no
reference model) and there won't be an available reference model if no
other models are opened, but the segmentation cursor still spawns.
Previously, the tool would still try to segment nothing and report a
traceback to the user when it failed. Now, it fails gracefully.