refactor(segmentations): Move triggers to their own file
The plane viewers needed a way to react to the segmentations command
being issued and redraw themselves if that command modified a
segmentation. Rather than adding yet another layer of
interconnectedness, something this module is already trying to move away
from, there is now a file: segmentations/triggers.py, which defines and
tracks triggers for this entire module.