refactor(segmentations): Enable GUIDELINES_VISIBILITY_CHANGED trigger
Currently, coordinating all the different ways the UI has to react to
changes is a cumbersome, fragile process wherein objects have to have
direct knowledge of each other. This refactor addresses one part of
that: making sure all the PlaneViewers and the Segmentation tool react
to changing the visibility of the slice guidelines without knowing about
each other by making this operating a trigger on changing a hidden setting.