fix(core): Work around setuptools#4170
There is a bug in importlib.metadata that makes it return duplicates of
installed packages when a package has been installed in editable mode.
This change works around this issue in ChimeraX by skipping bundle info
instances if we've already seen a bundle with the same name.

Workaround to Trac#15492