feat(rest_server): Optionally print output to ChimeraX's log
It's a problem that remotecontrol doesn't print status messages to
ChimeraX's log as well as the REST log. Because that information is lost
in ChimeraX, it would be impossible to get a sensible bug report out of
a restclient user. Now, the REST server can optionally be configured to
log to both the ChimeraX log and the client log.