feat(std_commands.graphics): Add checkUniforms to graphics command
Shader compilers are pretty aggressive about optimizing away uniforms if
it can be proven that they don't contribute to the final image. So when
a programmer tries to output, say, a box colored by some variable, it's
highly likely some uniform will be optimized out of the shader and when
the code for their drawing goes to set it, they'll get an OpenGLError.

These errors are totally recoverable, they just exist to ensure
unnecessary work isn't being done, so this behavior should be toggleable
for development purposes.