fix(segmentations): Hide the close button for orthoplane pseudomodels
The segmentations orthoplanes UI creates models that it manages outside
of the normal ChimeraX scene graph. It also puts panels in the volume
viewer that are associated with those models so that the user can change
the colors and contrast in the orthoplanes view. Those panels had
minimal customization but left the close button intact.

The user can accidentally click one of those close buttons and get
tracebacks, which also breaks that orthoplane. This commit removes
the close button, leaving it entirely up to the orthoplanes UI to
open and close those models.

Fixes Trac#15695