fix(core): Ensure app_bin_dir is the same for Python and ChimeraX launches on macOS
We try to set app_bin_dir to the root directory of the executable used
to launch ChimeraX. We use the realpath on all platforms to solve an issue
where the Linux version of ChimeraX may be launched by a symbolic link
to the real ChimeraX binary.

ChimeraX can also be launched by invoking python3 -m chimerax.core.
On macOS, the python that lives next to ChimeraX is symlinked to the
actual python executable deep in the framework directory. This change
makes it so that in that case we adjust the app_bin_dir accordingly.