fix(pre-commit): Don't error out on short paths
The pre-commit hook looks at path.split('/') to see if a
change is in a bundle, which can throw index error if the
path is too short/near the root. This change catches
the error and returns "" instead. It could return any string,
because what matters is we compare it against a generated list
of all bundle directory names.