[refactor] Bundle Triggers
Bundle triggers are moved out of the scene manager and are now stored in a specific module for the bundle called triggers.py.

[docs] Bundle Triggers

Documenting bundle defined triggers.

[feature] scene EDITED trigger

Added EDITED trigger to bundle triggers for when scenes are edited.

[fix] Trigger in manager.py

Trigger in manager.py was still referring to class instance attribute for triggers instead of bundle trigger.