[feature] SceneItem widget
Implementation for scenes that are added into the grid layout in the scenes widget.

[change] Remove SceneItem onClick

Remove the scaffolding for onClick from the SceneItem widget.

[change] SceneItem fixed text image pos

Fix the relative position of the text label and the thumbnail image in the SceneItem widget.

[change] SceneItem label alignment

Fix the label in the SceneItem having an offset center alighment.