[feature] ScenesWidget Grid Resizing
The scenes widget grid resizes when the window changes size. This is implemented with an update layout method in the widget that is called on the resize event override implementation.

[change] ScenesWidget grid resizing update_layout

The update_layout method calculates an appropriate minimum height for the grid layout once the rows and columns are filled in. This makes sure that on the startup of the tool the rows are not smashed together.

[change] Remove redundant size policy

Remove redundant size policy out of the constructor for the ScenesWidget.

[docs] ScenesWidget update_layout method comment

[refactor] ScenesWidget ITEM_WIDTH constant

There is a constant in the ScenesWidget class that represents how wide each grid space should be.