[feature] Scene Edited trigger tool handler
Implement a handler for the EDITED trigger from bundle triggers in the tool. The callback gets the name of the adjusted scene as it's data value. From there the corresponding SceneItem widget is targeted with a thumbnail update from the new thumbnail data from the scene in the session's scene manager.