[feature] Show latest scene first
Any scene that is edited gets moved to the top of the list so that the list maintains chronological order of latest edited scene. Uses new implementation of set_latest_scene on the ScenesWidget which takes a scene name and moves that scene to the start of the grid should it exist.