[fix] Column adjustments on resize
Start of work to correctly adjust column spacing for resizing events.

[fix] Column adjustments on resize

Implement custom SceneScrollArea which can track the resizing events of the view port. The SceneScrollArea also contains a grid widget which preforms similar functions to the old ScenesWidget. With all the functionality of the QScrollArea and the Scenes Widget inside the same widget, it is much easier to sizing relative to each widget.

[fix] Column adjustments on resize

Move the rest of the functionality from ScenesWidget into the SceneScrollArea widget. Update the callbacks in the tool class to reflect the methods getting moved into SceneScrollArea.