[feature] SceneItem Highlighting
[feature] Highlight Scene Trigger

Highlight scene trigger can be used in the tool to indicate when a SceneItem widget is clicked on and needs to be highlighted. Highlighting a scene is useful in being able to target edit and delete button actions.

[feature] Highlight Scene On Click

When a SceneItem widget is clicked on, it applies a border styling to itself.

[feature] Track Highlighted Scene

The main tool has a field that tracks the highlighted scene. If the highlighted scene changes it will un-highlight the old scene. In addition, if a highlighted scene is selected if it is already highlighted, the scene restore command is run.

[feature] Edit Highlighted Scene

Clicking the edit button while there is a scene highlighted targets the edit command at the highlighted scene.