[feature] Tool Delete Scene Button
Connect removing a scene item widget to the clicked signal of the delete button. The connected method will run the delete command and if no scene is highlighted will warn the user that they need to highlight a scene to delete. The rest of deleting is handled by the DELETE trigger callback handler.