[change] scene.py md_scene_implementation None Return
The md_scene_implementation function in scene.py required that Model implemented scenes. To remove this dependency, instead of the function default returning model when no other Scene support is found, the function returns None. When it's time to collect Scene snapshots, if the md_scene_implementation returns None, the model is skipped.

[docs] scene.py md_scene_implementation None Return