[feature] Snapshot Implementation in Tool
[feature] Snapshot Implementation in Tool

The tool implements snapshot take and restore. Add saving the active highlighted scene in the scroll menu.

[feature] SceneItem Grid Order in Snapshot

Add ordering of the SceneItem grid in the scroll area to the tools snapshot take and restore. Implement a take_snapshot and set_state_from_snapshot in the SceneScrollArea class to handle recording and resetting the order of the grid. Those methods are called on the scroll area through the ToolInstances snapshot methods.