[test] Initial pytests For Scenes Bundle
[test] Start Testing init_from_session in Scene Class.

[test] Scene initialization test.

Test that all the attributes that appear in the take_snapshot method are initialized correctly when a scene is initialized from the session.

[test] conftest.py test_production_session Patch Scene.take_thumbnail()

The take_thumbnail() of Scene will error in no gui mode. Patch the method with a placeholder string when configuring the test_production_sesison.

[test] Add manager_test.py to Test the Scene Manager

The first test that is added is for the scene_exists method in the scene_manager.

[test] Replace pytest-mocker Usage With unittest.mock.patch.

Instead of requiring pytest-mock to be installed use unittest implementation to patch the implementation of take_thumbnail in test session fixture.

[test] Remove take_thumbnail Patch in scene_test.py

The take_thumbnail patch is part of hte test_production_session fixture so does not need to be here.

[test] Manager.delete_scene

Test that the delete_scene method removes a created scene and triggers the DELETE trigger properly.