[test] setup_trigger_handler Function to Test Scene Trigger Activations
The setup_trigger_handler function creates a generic trigger handler. The function takes a trigger_type and expected_scene_name param to set up the handler. When the handler is called it asserts that the appropriate trigger type and scene name were passed. It then adjusts it's attribute 'called' to indicate if the handler was called.