feat(graphics,map): Add a method to set shader options to the Drawing class, use it in Image3D
This commit adds a set_shader_options(self, renderer) method to the
Drawing class, which is called right after shader activation, to set any
uniforms that are unique to that drawing. Usually shader uniform values
are set in the renderer and cached there, and doing things this way for
VolumeRaycastDrawing was also discussed.

This commit fixes an issue that prevented VolumeRaycastDrawings from
being drawn when any other drawing was in the scene, since the raycast
drawing would attempt to set its maximum and current extent uniforms
before its shader was even active.