fix(dicom): Ignore the rotations of DICOM datasets
DICOM datasets can be rotated in all kinds of wonky ways, but IMO the
important things are the pixel and slice spacings, since those variables
make sure DICOM datasets appear correctly proportioned. Ignoring
rotations means that we don't have to figure out how to rotate the data
or move the camera in the raycasting shader, and the PlaneViewers don't
have to calculate relative-to-dataset axis down which to point their
cameras. If someone complains we can explore those solutions but for now
we won't.