fix(graphics): Truncate float64 textures to float32
I made this change while working on a NIfTI file that was breaking
raycasting. The underlying image is a binary segmentation. While
the texture is correctly loaded for the normal planes drawings, for
whatever reason, when requesting a raycast image we get values
close to 0 and close to 1 in float64 format instead of 0s and 1s.
Float64 isn't a texture format supported by OpenGL, so this commit
truncates float64 to float32.