fix(segmentations): Set the projection mode of orthoplane volumes to '3d'
This commit fixes an issue where if the user tried to open the
orthoplane viewers while the original volume's projection mode
was set to rays, the orthoplanes viewers' derived volumes would
be invisible. This happened because those volumes would inherit
the original volume's projection mode, and the rays projection
mode seldom if ever works for a single slice.