feat: sync RatingDelegate editor with model after de-selecting
Alter the setModelData implementation in the RatingDelegate to adjust the actual AtomicStructure that is in the row instead of only the QAbstractItemModel. The paint method only gets called with the ItemTable's data_fetch. The ItemTable implementation of data_fetch looks at the AtomicStructure attributes instead of the QAbstractItemModel. This syncs the RatingColumn combobox rendering with the editor menu.