refactor: viewdock col refresh with MODEL_DISPLAY_CHANGED trigger
Updating display column on 'hide/show all' button click does not handle display changes from cl. Replace updating the full column with a trigger handler for the MODEL_DISPLAY_CHANGED trigger. The handler calls update cell for the display cell for the model that was changed. This is technically redundant when selecting and deselecting in the table.