refactor(viewdock): Description box generated from col data_fetch
Instead of using the model attributes to generate the model description box, use the cols and their data_fetch's. This ensures that if there is a column that uses a custom data_fetch, it's info will be displayed the same way in the description box as it is in the table. Data in the viewdock attr of docking models that does not have a data_fetch through the table will have their value displayed without any adjustment.