fix(viewdock): Overlayed text in description box when table right-click
When right-clicking the table, the row is selected, but the context menu is also pulled up, which takes over the QT event loop. In this case, old label widgets in the description box from the previous selection hang as queued for deletion and the new labels widgets get displayed over them. Clear the text of the old labels before adding new widgets so that they are invisible while queued for deletion.