fix(sideview): Don't set the parent to None before deleting anymore
We used to set the parent of the side view graphics window to None
when we used PyQt5, since not taking that action caused an already
deleted widget to receive a resize event which would crash ChimeraX
(see Trac#2050). Now, with PyQt6 or PySide6, the already deleted widget
receives a parentChanged event! In newer versions of Qt, using either
wrapper library, this again causes ChimeraX to crash.

Closes Trac#18474