fix(map): Reenable OpenMP
There are 4 situations to handle when distributing OpenMP-enabled code:

1) Linux
2) Flatpaks
3) Windows
4) macOS

Windows is actually the simplest platform. The common runtime we ship
with ChimeraX already has OpenMP code.

On macOS, we can ship the Homebrew OpenMP with the bundle that uses it,
or put it in ChimeraX.app/Contents/lib if more than one comes to
depend on it. The changes to bundle builder smartly detect whether
OpenMP is actually installed on the builder's system, and if it
is not installed the build simply becomes single-threaded.

On Linux, we can rely on the fact that we build against Rocky 8 and
trust OpenMP's backwards compatibility, the way we do for libc.

For the Flatpaks, we can use the Flatpak runtime's OpenMP.