feat(graphics): Projection-independent raycasting
Previously, raycasting depended on the camera's position and FOV,
which locked it to the mono camera. This code reconstructs rays
from inverse projection and view matrices instead, rendering (NPI)
the code independent of the camera type. Raycasting can now be used
with ortho, stereo, 360 cameras, etc.

The jitter has also been slightly reworked to improve the appearance
of volumes and reduce banding.